refactor(persona): split evaluator.rs (1231 LOC) into focused submodules (#1208)
`persona/evaluator.rs` was a single 1231-LOC file mixing four
independent concerns: persona sleep state, per-room rate limiting,
post-inference adequacy check, and the `full_evaluate` gate
orchestrator. Split into:

  evaluator/mod.rs         (886 LOC) — gate orchestrator, FullEvaluate
                                       request/result types, 18 gate
                                       integration tests
  evaluator/sleep_state.rs (99 LOC)  — SleepMode + SleepState +
                                       2 unit tests
  evaluator/rate_limiter.rs (132 LOC) — RateLimiterState + RoomRateState
                                        + 2 unit tests (track_response,
                                        rate_limit_expired)
  evaluator/adequacy.rs    (207 LOC) — RecentResponse + AdequacyResult +
                                       check_response_adequacy + 7 tests

`persona/mod.rs` re-exports unchanged: `pub use evaluator::{...}` still
exposes SleepMode, SleepState, RateLimiterState, AdequacyResult,
RecentResponse, GateDetails, FullEvaluateRequest, FullEvaluateResult.
External callers see no API change.

Why these specific cuts:
- SleepState is reused independently anywhere a persona's voluntary
  attention state matters (not just Gate 1).
- RateLimiterState is a per-room cadence tracker that's a SIGNAL to the
  LLM, not a hard gate — independent of full_evaluate.
- Adequacy check is a separate phase (post-inference, not pre-response)
  that happens to share the file because it was written together.

Tests:
- `cargo test --features metal,accelerate persona::evaluator` →
  32 passed (every test from the original file, redistributed by domain).
- Full `persona::*` test suite → 451 passed, 0 failed, 3 ignored
  (no other module's imports broke).

Each new file < 250 LOC, mod.rs < 1000 LOC. Closes #1208 for the
`evaluator.rs` slice; `admission.rs` (1225) and `model_resolver.rs`
(1232) remain as separate cards.
